Contains most experiments from wandb

+++ model.py
@@ -212,8 +212,8 @@ class SkynetModel(nn.Module):
         offset = torch.empty((distance.shape))
         frequency[features[:,7] == 1] = torch.tensor(2.63)
         frequency[features[:,7] != 1] = torch.tensor(0.811)
-        offset[features[:,7] == 1] = torch.tensor(self.offset_2630)
-        offset[features[:,7] != 1] = torch.tensor(self.offset_811)
+        offset[features[:,7] == 1] = torch.tensor(self.offset_2630).float()
+        offset[features[:,7] != 1] = torch.tensor(self.offset_811).float()
         return frequency, offset
 
     def MSE_physicsmodel(self, distance, targets):
